started data cleaning and feature extraction
Got a little stuck while extracting a subtype column. Went back to the data to find that the notes field was not what I expected, so I opened up the app on my phone and figured out how to export the data from there, which includes the notes I wanted. I'm committing what I had for posterity before overwriting it.